moving add owner to channel to separate file
